pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-infrastructure-pyproject-versioning"
version = "0.0.1a2"
description = "Infrastructure layer for pythoneda-artifact/pyproject-versioning"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-infrastructure/pyproject-versioning"
packages = [{include = "pythonedaartifactinfrastructurepyprojectversioning"}]

[tool.poetry.dependencies]
dbus-next = "^0.2.3"
dulwich = "^0.21.5"
grpcio = "^1.41.0"
paramiko = "^2.11.0"
python = "^3.8"
pythoneda-artifact-event-versioning = "^0.0.1a1"
pythoneda-artifact-event-infrastructure-versioning = "^0.0.1a1"
pythoneda-artifact-pyproject-versioning = "^0.0.1a2"
pythoneda-base = "^0.0.1a16"
pythoneda-infrastructure-base = "^0.0.1a12"
pythoneda-shared-git = "^0.0.1a3"
requests = "^2.28.1"
semver = "^3.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
